fix tabs destroyed if switched away during first webview attach
Fix #13798 since "display: none" and then "display: ''" of a webview will cause the guest to (asynchronously) detach and attach again, and by that time we had attached another guest, which causes the original tab to be destroyed since we are attaching a guest to a webview which already has a guest.
Luckily, it doesn't look like this workaround is needed in order to paint the first attachGuest of a webview anymore.